Move ZingTouch entry point up into index.js
Previously, index.js wasn't so much providing an entry point as just
requiring and then re-exporting the ZingTouch.js file. This didn't make
any sense to me. Maybe some day I'll find out why this was done.
